requirements.txt
pillow==10.3.0
torchvision==0.21.0
torch==2.6.0